add ad

+++ app/controllers/categories_controller.rb
@@ -7,7 +7,7 @@ class CategoriesController < ApplicationController
   
   def show
     @category = Category.find(params[:id])
-    @posts_page = @category.posts.order("updated_at DESC").page(params[:page]).per(7)
+    @posts_page = @category.posts.order("updated_at DESC").page(params[:page]).per(6)
   end
   
   def new
@@ -17,11 +17,16 @@ class CategoriesController < ApplicationController
   def create
     @category = Category.new(cat_params)
 
-    if @category.save
-      redirect_to @category, notice: "Category was saved successfully."
-    else
-      flash.now[:alert] = error_messages(@category)
-      render :new
+    if !is_admin?
+      flash[:alert] = "Only admin can create category."
+      redirect_to @category
+    else      
+      if @category.save
+        redirect_to @category, notice: "Category was saved successfully."
+      else
+        flash.now[:alert] = error_messages(@category)
+        render :new
+      end
     end
   end
   
@@ -33,24 +38,34 @@ class CategoriesController < ApplicationController
     @category = Category.find(params[:id])
     @category.assign_attributes(cat_params)
     
-    if @category.save
-      flash[:notice] = "Category was updated successfully."
+    if !is_admin?
+      flash[:alert] = "Only admin can edit category."
       redirect_to @category
-    else
-      flash.now[:alert] = error_messages(@category)
-      render :edit
+    else      
+      if @category.save
+        flash[:notice] = "Category was updated successfully."
+        redirect_to @category
+      else
+        flash.now[:alert] = error_messages(@category)
+        render :edit
+      end
     end
   end
   
   def destroy
     @category = Category.find(params[:id])
     
-    if @category.destroy
-      flash[:notice] = "\"#{@category.name}\" was deleted successfully."
-      redirect_to action: :index
-    else
-      flash.now[:alert] = "There was an error deleting the category."
-      render :show
+    if !is_admin?
+      flash[:alert] = "Only admin can delete category."
+      redirect_to @category
+    else      
+      if @category.destroy
+        flash[:notice] = "\"#{@category.name}\" was deleted successfully."
+        redirect_to action: :index
+      else
+        flash.now[:alert] = "There was an error deleting the category."
+        render :show
+      end
     end
   end
